requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
aiohttp==3.9.1
feedparser==6.0.10
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.6 